+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.86.0] - 2020-11-17
+
 ### Changed
 - `enableFiltersFetchOptimization` now working on mobile.